Disable noalias propagation for HLSL <= 2021
HLSL has very few ways that users can express pointers in code, as a
result propagating alias information during inlining costs a lot of
compile time with minimal benefit to generated code.

This change adds a new mechanism for putting metadata into the IR
module, to feed things through to the middle-end, and removing the
metadata at the end of optimization.