Remove `UInt32Add` and `Int32ToUInt32`
All call sites of `UInt32Add` except one are adding one to a value
converted from a signed integer of the same size. As a result they
cannot overflow.

The one remaining case is a place where overflow could occur, but it is
easy enough to check for overflow and return an error.

All call sites of `Int32ToUInt32` are on the result of
`WideCharToMultiByte` which is defined to return `0` on error or a
positive number on success. Given that definition, it is always safe to
cast the signed integer to an unsigned of the same size.

There is one additional case I added an overflow check in this code. In
`CodePageBufferToUtf8` we convert a signed 32-bit integer to an
unsigned 32-bit integer then increment the unsiged value and cast it
back to a signed value. This can overflow, and should be handled
appropriately.