Experimental reimplementation of `out` and `inout`
This change rewrites `out` and `inout` parameters to be implemented on
the AST. The parameter copying behavior is modeled on the AST and
generated through codegen.

Instead of using the previously existing infrastructure for generating
HLSL's copy-in/copy-out semantaics, this change generates temporaries
based on the AST, uses AST-based analysis for optimziation, and uses
the Objective-C indirect writeback support for copy-out.