[Draft] Rewrite output parameters
This change is a huge swath of changes rewriting how HLSL represents
pointers, references and arrays as well as the handling of output
parameters in function signatures.

The main changes included in this PR are:
* Restoring Array->Pointer Decay in the HLSL ASTs.
* Representing all output parameters using reference types in the AST.
* Adding a new HLSLOutParamExpr AST node for output parameters.
*Adding a new HLSLArrayTemporaryExpr to represent temporary array
values.

There are also a large number of associated fixes and patches to make
the code generation and optimization code resilient to the new AST and
IR structures that are generated.